Split classes into individual files
+ some docs
