+++ OpenGpxTracker/DefaultDateFormat.swift
@@ -62,19 +62,3 @@ class DefaultDateFormat {
     }
 
 }
-// from: https://stackoverflow.com/a/45073012/13292870
-// count occurances
-extension String {
-    /// stringToFind must be at least 1 character.
-    func countInstances(of stringToFind: String) -> Int {
-        assert(!stringToFind.isEmpty)
-        var count = 0
-        var searchRange: Range<String.Index>?
-        while let foundRange = range(of: stringToFind, options: [.literal], range: searchRange) {
-            count += 1
-            searchRange = Range(uncheckedBounds: (lower: foundRange.upperBound, upper: endIndex))
-        }
-        return count
-    }
-    
-}
