+++ OpenGpxTracker/UIColor+DarkMode.swift
@@ -46,11 +46,3 @@ extension UIColor {
     }
     
 }
-
-extension UIColor {
-    static let lightKeyboard = UIColor(red: 209/255, green: 213/255, blue: 219/255, alpha: 1.00)
-    static let darkKeyboard = UIColor(red: 36/255, green: 36/255, blue: 36/255, alpha: 1.00)
-    
-    static let highlightLightKeyboard = UIColor(red: 229/255, green: 233/255, blue: 239/255, alpha: 1.00)
-    static let highlightDarkKeyboard = UIColor(red: 56/255, green: 56/255, blue: 56/255, alpha: 1.00)
-}
