+++ OpenGpxTracker/UIColor+Keyboard.swift
@@ -1,0 +1,21 @@
+//
+//  UIColor+Keyboard.swift
+//  OpenGpxTracker
+//
+//  Created by Vincent Neo on 17/4/20.
+//
+
+import UIKit
+
+/// For `DateFieldTypeView`. Meant to match default keyboard colors.
+extension UIColor {
+    /// Default Light Appearance Keyboard
+    static let lightKeyboard = UIColor(red: 209/255, green: 213/255, blue: 219/255, alpha: 1.00)
+    /// Default Dark Appearance Keyboard
+    static let darkKeyboard = UIColor(red: 36/255, green: 36/255, blue: 36/255, alpha: 1.00)
+    
+    /// Button Highlight Light Appearance Keyboard
+    static let highlightLightKeyboard = UIColor(red: 229/255, green: 233/255, blue: 239/255, alpha: 1.00)
+    /// Button Highlight Dark Appearance Keyboard
+    static let highlightDarkKeyboard = UIColor(red: 56/255, green: 56/255, blue: 56/255, alpha: 1.00)
+}
